+++ CHANGES.rst
@@ -1,3 +1,6 @@
+**v1.5** - 
+
+
 **v1.2** 
 
 * Create backup of output file when it exists and is non-empty